chore(rudder-sdk-react-native): rename Util file to FiterNaN
