+++ libs/sdk/src/FilterNaN.ts
